refactor: move parseErrorMsg to utils(bc of pure func)
