+++ src/lib/bookingFormUtils.ts
@@ -7,12 +7,13 @@
  * @contact intra: @xifeng
  */
 
+import { AxiosError } from 'axios';
 import { addMinutes, differenceInMinutes, isAfter, isBefore, isEqual } from 'date-fns';
 
-import type { FormProp, FormType } from '@/components/bookingForm/BookingForm';
 import type { Slot } from '@/components/bookingForm/ScrollSlotPicker';
 import { OPEN_HOURS_IDX, ROOM_MAP, TIME_SLOT_INTERVAL } from '@/config';
 import { ThrowInternalError } from '@/lib/errorHandler';
+import type { FormProp, FormType } from '@/lib/hooks/useBookingForm';
 import type { DayBookings } from '@/lib/weekBookings';
 
 import type { BookingFromApi, UpsertBooking } from './schema';
@@ -133,4 +134,14 @@ const overlappingCheck = (start: string, end: string, endSlots: Slot[]): boolean
   return true; // should not be here.
 };
 
-export { calculateSlots, initForm, overlappingCheck };
+const parseErrorMsg = (error: unknown): string => {
+  if (error instanceof AxiosError) {
+    return error.response?.data?.message ?? error.message ?? 'Server responded with an error.';
+  }
+  if (error instanceof Error) {
+    return error.message;
+  }
+  return 'Unknown error occurred.';
+};
+
+export { calculateSlots, initForm, overlappingCheck, parseErrorMsg };
